+++ plugins/woocommerce-admin/packages/components/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 -   Fix usage of Wordpress DatePicker component in `DatePicker`. #7982
 -   Fix select-control component label/value alignment. #8045
+-   Fix clicking the error message opens the dropdown. #8094
 
 # 8.1.1